Refactor community page logic into new component
Moved the main logic of the community page from src/app/community/page.tsx to a new reusable CommunityPageComponent in src/components/views/community/community-page-component.tsx. Updated usage in dashboard/following-objects.tsx to use the new component, improving code organization and reusability.
