+++ src/components/views/dashboard/following-objects.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import CommunityPage from '@/app/community/page'
 import React from 'react'
 import { useAccount } from '@/libs/utils/account'
 import { AccountStatus } from '@/libs/constants/account.constant'
+import { CommunityPageComponent } from '@/components/views/community/community-page-component'
 
 
 export const FollowingObjects = () => {
@@ -12,7 +12,7 @@ export const FollowingObjects = () => {
     return (
         <div className='flex flex-col gap-y-3'>
             <h2 className='text-24-bold text-tunelog-light'>From Followings</h2>
-            <CommunityPage viewOnlyFollowing/>
+            <CommunityPageComponent viewOnlyFollowing/>
         </div>
     )
 }
